pyproject.toml
[project]
name = "multizone-app"
version = "0.1.0"
description = "A Streamlit application featuring a multi-zone interactive grid."
authors = [{name = "Your Name", email = "youremail@example.com"}]
license = {text = "MIT"}
dependencies = [
    "streamlit>=1.28.0",
    "Pillow>=9.0.0",
    "pyserial>=3.5"
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"